+++ .github/linters/.eslintrc.yml
@@ -1,8 +1,7 @@
 env:
-  commonjs: true
+  node: true
   es6: true
   jest: true
-  node: true
 
 globals:
   Atomics: readonly
@@ -15,24 +14,22 @@ ignorePatterns:
   - '**/coverage/.*'
   - '*.json'
 
-parser: '@babel/eslint-parser'
+parser: '@typescript-eslint/parser'
 
 parserOptions:
   ecmaVersion: 2023
   sourceType: module
-  requireConfigFile: false
-  babelOptions:
-    babelrc: false
-    configFile: false
-    presets:
-      - jest
+  project:
+    - './.github/linters/tsconfig.json'
+    - './tsconfig.json'
 
 plugins:
   - jest
+  - '@typescript-eslint'
 
 extends:
   - eslint:recommended
-  - plugin:github/recommended
+  - plugin:@typescript-eslint/recommended-type-checked
   - plugin:jest/recommended
 
 rules:
@@ -41,10 +38,27 @@ rules:
     'eslint-comments/no-use': 'off',
     'eslint-comments/no-unused-disable': 'off',
     'i18n-text/no-en': 'off',
-    'import/no-commonjs': 'off',
     'import/no-namespace': 'off',
     'no-console': 'off',
-    'no-unused-vars': 'off',
-    'prettier/prettier': 'error',
-    'semi': 'off'
+    'semi': 'off',
+    '@typescript-eslint/array-type': 'error',
+    '@typescript-eslint/consistent-type-assertions': 'error',
+    '@typescript-eslint/explicit-member-accessibility':
+      ['error', { 'accessibility': 'no-public' }],
+    '@typescript-eslint/explicit-function-return-type':
+      ['error', { 'allowExpressions': true }],
+    '@typescript-eslint/no-empty-interface': 'error',
+    '@typescript-eslint/no-extraneous-class': 'error',
+    '@typescript-eslint/no-inferrable-types': 'error',
+    '@typescript-eslint/no-non-null-assertion': 'warn',
+    '@typescript-eslint/no-unnecessary-qualifier': 'error',
+    '@typescript-eslint/no-useless-constructor': 'error',
+    '@typescript-eslint/no-var-requires': 'error',
+    '@typescript-eslint/prefer-for-of': 'warn',
+    '@typescript-eslint/prefer-function-type': 'warn',
+    '@typescript-eslint/prefer-includes': 'error',
+    '@typescript-eslint/prefer-string-starts-ends-with': 'error',
+    '@typescript-eslint/promise-function-async': 'error',
+    '@typescript-eslint/require-array-sort-compare': 'error',
+    '@typescript-eslint/space-before-function-paren': 'off'
   }
