+++ package.json
@@ -12,15 +12,15 @@
     "url": "https://github.com/chrisreddington/devcontainer-validator/issues"
   },
   "keywords": [
-    "GitHub",
-    "Actions",
-    "JavaScript"
+    "actions",
+    "node",
+    "setup"
   ],
   "exports": {
     ".": "./dist/index.js"
   },
   "engines": {
-    "node": ">=22"
+    "node": ">=21"
   },
   "scripts": {
     "bundle": "npm run format:write && npm run package",
@@ -29,29 +29,32 @@
     "format:write": "npx prettier --write .",
     "format:check": "npx prettier --check .",
     "lint": "npx eslint . -c ./.github/linters/.eslintrc.yml",
-    "package": "npx ncc build src/index.js -o dist --source-map --license licenses.txt",
+    "local-action": "npx local-action . src/main.ts .env",
+    "package": "npx ncc build src/index.ts -o dist --source-map --license licenses.txt",
     "package:watch": "npm run package -- --watch",
     "test": "npx jest",
     "all": "npm run format:write && npm run lint && npm run test && npm run coverage && npm run package"
   },
   "license": "MIT",
-  "eslintConfig": {
-    "extends": "./.github/linters/.eslintrc.yml"
-  },
   "jest": {
+    "preset": "ts-jest",
     "verbose": true,
     "clearMocks": true,
     "testEnvironment": "node",
     "moduleFileExtensions": [
-      "js"
+      "js",
+      "ts"
     ],
     "testMatch": [
-      "**/*.test.js"
+      "**/*.test.ts"
     ],
     "testPathIgnorePatterns": [
       "/node_modules/",
       "/dist/"
     ],
+    "transform": {
+      "^.+\\.ts$": "ts-jest"
+    },
     "coverageReporters": [
       "json-summary",
       "text",
@@ -66,17 +69,22 @@
     "@actions/core": "^1.11.1"
   },
   "devDependencies": {
-    "@babel/core": "^7.26.0",
-    "@babel/eslint-parser": "^7.25.9",
-    "@babel/preset-env": "^7.26.0",
-    "@github/local-action": "^2.1.2",
+    "@github/local-action": "^2.1.3",
+    "@jest/globals": "^29.7.0",
+    "@types/jest": "^29.5.14",
+    "@types/node": "^22.8.7",
+    "@typescript-eslint/eslint-plugin": "^8.12.2",
+    "@typescript-eslint/parser": "^8.12.2",
     "@vercel/ncc": "^0.38.2",
-    "babel-preset-jest": "^29.6.3",
-    "eslint": "^8.57.1",
-    "eslint-plugin-github": "^5.0.2",
+    "eslint": "^8.57.0",
     "eslint-plugin-jest": "^28.8.3",
+    "eslint-plugin-jsonc": "^2.16.0",
+    "eslint-plugin-prettier": "^5.2.1",
     "jest": "^29.7.0",
     "make-coverage-badge": "^1.2.0",
-    "prettier": "^3.3.3"
+    "prettier": "^3.3.3",
+    "prettier-eslint": "^16.3.0",
+    "ts-jest": "^29.2.5",
+    "typescript": "^5.6.3"
   }
 }
