--- src/index.js
@@ -1,6 +1,0 @@
-/**
- * The entrypoint for the action.
- */
-const { run } = require('./main')
-
-run()
