+++ src/index.ts
@@ -1,0 +1,4 @@
+import { run } from './main'
+
+// eslint-disable-next-line @typescript-eslint/no-floating-promises
+run()
